+++ src/Classes/ServiceAPI/MyRadio_Swagger.php
@@ -13,7 +13,7 @@ use \ReflectionException;
 
 use \MyRadio\Config;
 use \MyRadio\Database;
-use \MyRadio\MyRadio\CoreUtils;
+use \MyRadio\MyRadio\AuthUtils;
 
 /**
  * The Swagger class is an Implementation of https://developers.helloreverb.com/swagger/
@@ -263,7 +263,7 @@ class MyRadio_Swagger
         } else {
             $long_desc .= '<br>The following permissions enable access to this method:';
             foreach ($requirements as $typeid) {
-                $long_desc .= '<br> - ' . CoreUtils::getAuthDescription($typeid);
+                $long_desc .= '<br> - ' . AuthUtils::getAuthDescription($typeid);
             }
         }
 
@@ -300,7 +300,7 @@ class MyRadio_Swagger
 
     /**
      * Return the methods this endpoint allows.
-     * 
+     *
      * Specify these with one or more @api decorators.
      * Defaults to GET only.
      * Defaults to POST if the method begins with 'set' (e.g. setIntro)
