+++ src/Controllers/MyRadio/a-timeslotSignin.php
@@ -6,12 +6,13 @@
  * @package MyRadio_Core
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_Timeslot;
 
 $ts = MyRadio_Timeslot::getInstance($_REQUEST['timeslotid']);
 if ($ts->getSeason()->getShow()->isCurrentUserAnOwner()
-    or CoreUtils::hasPermission(AUTH_EDITSHOWS)
+    || AuthUtils::hasPermission(AUTH_EDITSHOWS)
 ) {
     $data = $ts->getSigninInfo();
     CoreUtils::dataToJSON($data);
