+++ src/Controllers/MyRadio/permissionUsage.php
@@ -6,11 +6,12 @@
  * @package MyRadio_Core
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 
-$usage = CoreUtils::getAuthUsage($_REQUEST['typeid']);
+$usage = AuthUtils::getAuthUsage($_REQUEST['typeid']);
 
 CoreUtils::getTemplateObject()->setTemplate('MyRadio/permissionUsage.twig')
-        ->addVariable('title', 'Permission Usage | '.CoreUtils::getAuthDescription($_REQUEST['typeid']))
+        ->addVariable('title', 'Permission Usage | '.AuthUtils::getAuthDescription($_REQUEST['typeid']))
         ->addVariable('usage', $usage)
         ->render();
