+++ src/Controllers/MyRadio/timeslot.php
@@ -8,6 +8,7 @@
  */
 
 use \MyRadio\Config;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_User;
 use \MyRadio\ServiceAPI\MyRadio_Timeslot;
@@ -21,7 +22,7 @@ function setupTimeslot($timeslot)
     }
 
     //Can the user access this timeslot?
-    if (!($timeslot->getSeason()->getShow()->isCurrentUserAnOwner() || CoreUtils::hasPermission(AUTH_EDITSHOWS))) {
+    if (!($timeslot->getSeason()->getShow()->isCurrentUserAnOwner() || AuthUtils::hasPermission(AUTH_EDITSHOWS))) {
         $message = "You don't have permission to view this show";
         require_once 'Controllers/Errors/403.php';
     } else {
@@ -42,7 +43,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
     } else {
         CoreUtils::backWithMessage("Cannot select empty timeslot");
     }
-} elseif (isset($_GET['current']) && $_GET['current'] && CoreUtils::hasPermission(AUTH_EDITSHOWS)) {
+} elseif (isset($_GET['current']) && $_GET['current'] && AuthUtils::hasPermission(AUTH_EDITSHOWS)) {
     //Submitted Current
     setupTimeslot(MyRadio_Timeslot::getCurrentTimeslot());
 
@@ -62,7 +63,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
      * People with AUTH_EDITSHOWS can see all timeslots here
      */
     $shows = MyRadio_User::getInstance()->getShows();
-    if (CoreUtils::hasPermission(AUTH_EDITSHOWS)) {
+    if (AuthUtils::hasPermission(AUTH_EDITSHOWS)) {
         if (isset($_GET['all'])) {
             $shows = MyRadio_Show::getAllShows();
             $twig->addVariable('allTimeslots', 'on');
