+++ src/Controllers/Podcast/editPodcast.php
@@ -5,6 +5,7 @@
  */
 
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_Podcast;
 use \MyRadio\ServiceAPI\MyRadio_Show;
@@ -29,7 +30,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
         // Check if user can edit this podcast
         if (!in_array($podcast->getID(), MyRadio_Podcast::getPodcastIDsAttachedToUser())) {
-            CoreUtils::requirePermission(AUTH_PODCASTANYSHOW);
+            AuthUtils::requirePermission(AUTH_PODCASTANYSHOW);
         }
 
         $podcast->setMeta('title', $data['title'])
@@ -62,7 +63,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
 
         // Check if user can edit this podcast
         if (!in_array($podcast->getID(), MyRadio_Podcast::getPodcastIDsAttachedToUser())) {
-            CoreUtils::requirePermission(AUTH_PODCASTANYSHOW);
+            AuthUtils::requirePermission(AUTH_PODCASTANYSHOW);
         }
 
         $podcast
