+++ src/Controllers/Profile/view.php
@@ -8,6 +8,7 @@
  * @package MyRadio_Profile
  */
 
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\MyRadio_User;
 use \MyRadio\ServiceAPI\MyRadio_TrainingStatus;
@@ -28,7 +29,7 @@ if ($user->isOfficer()) {
     $userData['email'] = $user->getPublicEmail();
 }
 
-if (CoreUtils::hasPermission(AUTH_VIEWOTHERMEMBERS)) {
+if (AuthUtils::hasPermission(AUTH_VIEWOTHERMEMBERS)) {
     $userData['email'] = $user->getEmail();
     $userData['eduroam'] = $user->getEduroam();
     $userData['local_alias'] = $user->getLocalAlias();
@@ -56,9 +57,9 @@ if ($user->getID() === $visitor->getID() or $visitor->hasAuth(AUTH_EDITANYPROFIL
         .'">Edit Profile</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_IMPERSONATE)
+if (AuthUtils::hasPermission(AUTH_IMPERSONATE)
     && ($user->hasAuth(AUTH_BLOCKIMPERSONATE) === false
-    or CoreUtils::hasPermission(AUTH_IMPERSONATE_BLOCKED_USERS))
+    || AuthUtils::hasPermission(AUTH_IMPERSONATE_BLOCKED_USERS))
 ) {
     $template->addVariable(
         'impersonateurl',
@@ -71,7 +72,7 @@ if (CoreUtils::hasPermission(AUTH_IMPERSONATE)
         .'">Impersonate User</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_LOCK)) {
+if (AuthUtils::hasPermission(AUTH_LOCK)) {
     $template->addVariable(
         'lockurl',
         '<a href="'
@@ -83,7 +84,7 @@ if (CoreUtils::hasPermission(AUTH_LOCK)) {
         .'">Disable Account</a>'
     );
 }
-if (CoreUtils::hasPermission(AUTH_MARKPAYMENT)) {
+if (AuthUtils::hasPermission(AUTH_MARKPAYMENT)) {
     $template->addVariable('can_mark_payments', true);
 }
 $template->render();
