+++ src/Controllers/Setup/dbdata.php
@@ -8,6 +8,7 @@
  */
 use \MyRadio\Database;
 use \MyRadio\MyRadioException;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 use \MyRadio\ServiceAPI\ServiceAPI;
 use \MyRadio\ServiceAPI\MyRadio_Officer;
@@ -46,7 +47,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         }
         foreach (json_decode(file_get_contents(SCHEMA_DIR . 'data-auth.json')) as $auth) {
             try {
-                CoreUtils::addPermission($auth[0], $auth[1]);
+                AuthUtils::addPermission($auth[0], $auth[1]);
             } catch (MyRadioException $e) {
                 $warnings[] = 'Failed to create Permission "' . $auth[0] . '". It may already exist.';
             }
@@ -55,7 +56,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $module = CoreUtils::getModuleId($actionauth[0]);
             $action = $actionauth[1] == null ? null : CoreUtils::getActionId($module, $actionauth[1]);
             $auth = $actionauth[2] == null ? null : constant($actionauth[2]);
-            CoreUtils::addActionPermission($module, $action, $auth);
+            AuthUtils::addActionPermission($module, $action, $auth);
         }
         foreach (json_decode(file_get_contents(SCHEMA_DIR . 'data-apiauth.json')) as $apiauth) {
             $db->query(
@@ -95,7 +96,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             //The getXxxId method creates an ID if they don't exist
             $module = CoreUtils::getModuleId($action[0]);
             $action = CoreUtils::getActionId($module, $action[1]);
-            CoreUtils::addActionPermission($module, $action, null);
+            AuthUtils::addActionPermission($module, $action, null);
         }
         break;
     case DBDATA_BLANK:
@@ -106,7 +107,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
         }
         foreach (json_decode(file_get_contents(SCHEMA_DIR . 'data-auth-min.json')) as $auth) {
             try {
-                CoreUtils::addPermission($auth[0], $auth[1]);
+                AuthUtils::addPermission($auth[0], $auth[1]);
             } catch (MyRadioException $e) {
                 $warnings[] = 'Failed to create Permission "'.$auth[0].'". It may already exist.';
             }
@@ -115,7 +116,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $module = CoreUtils::getModuleId($actionauth[0]);
             $action = $actionauth[1] == null ? null : CoreUtils::getActionId($module, $actionauth[1]);
             $auth = $actionauth[2] == null ? null : constant($actionauth[2]);
-            CoreUtils::addActionPermission($module, $action, $auth);
+            AuthUtils::addActionPermission($module, $action, $auth);
         }
         break;
     default:
