+++ src/Controllers/root_web.php
@@ -8,6 +8,7 @@
  */
 
 use \MyRadio\Config;
+use \MyRadio\MyRadio\AuthUtils;
 use \MyRadio\MyRadio\CoreUtils;
 
 require_once __DIR__.'/root.php';
@@ -52,7 +53,7 @@ if (isset($_REQUEST['request'])) {
  * IMPORTANT: This will cause a fatal error if an action does not have any permissions associated with it.
  * This is to prevent developers from forgetting to assign permissions to an action.
  */
-CoreUtils::requirePermissionAuto($module, $action);
+AuthUtils::requirePermissionAuto($module, $action);
 
 /**
  * If a Joyride is defined, start it
